# web/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Swarm Builder",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/settings": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get Settings",
        "description": "Everything the model-settings screen renders.\n\nReads the file fresh: a user can edit ``workspace/settings.json`` (or flip\nthe dry-run switch from another tab) between two requests.",
        "operationId": "get_settings_api_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Put Settings",
        "description": "Save the in-app model configuration and/or the dry-run switch.\n\nValidation happens before anything is written, so a rejected submission\nleaves the previous configuration exactly as it was. On success the\ncredential is published into this server's environment immediately (see\n:func:`swarm_builder.runtime.publish_secrets`), which is what makes the\nchange take effect for the next generate/compile/run without a restart.\n\nRaises:\n    HTTPException: 422 with every problem found, or 500 when the settings\n        file cannot be written (an unwritable workspace).",
        "operationId": "put_settings_api_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            }
          },
          "422": {
            "description": "the submitted provider/model/key combination is not usable"
          },
          "500": {
            "description": "the settings file could not be written"
          }
        }
      }
    },
    "/api/settings/test": {
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Test Connection",
        "description": "Make one minimal model call and report whether it worked.\n\nA brand-new user's first question is \"did I paste the key correctly?\", and\nthe honest answer requires a real call -- the keyless validation gate a\ncompile runs cannot answer it, because it injects a test model on purpose.\n\nEvery provider-side outcome (a refused key, a timeout, an unmappable\nroute, nothing configured at all) is a ``200`` with ``ok: false`` and a\nhuman-readable ``detail``; only a structurally invalid request is a\n``422``. Nothing here is persisted: testing an unsaved form must not\nchange the saved configuration.",
        "operationId": "test_connection_api_settings_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestConnectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestConnectionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResolvedDefaultOut": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "source": {
            "type": "string",
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "model",
          "source"
        ],
        "title": "ResolvedDefaultOut",
        "description": "The model a compile would use right now, plus where that came from."
      },
...
      "SavedModelOut": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "baseUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseurl"
          },
          "reasoningEffort": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoningeffort"
          },
          "hasApiKey": {
            "type": "boolean",
            "title": "Hasapikey"
          },
          "apiKeyHint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apikeyhint"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "model",
          "baseUrl",
          "reasoningEffort",
          "hasApiKey",
          "apiKeyHint"
        ],
        "title": "SavedModelOut",
        "description": "The saved model, minus the secret.\n\n``api_key_hint`` is the last four characters of the stored key -- enough\nfor a user to tell two keys apart, not enough to use one. The value itself\nis never part of any response body."
      },
...
      "SettingsResponse": {
        "properties": {
          "configPath": {
            "type": "string",
            "title": "Configpath"
          },
          "configError": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Configerror"
          },
          "model": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SavedModelOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "dryRun": {
            "type": "boolean",
            "title": "Dryrun"
          },
          "dryRunForcedByEnv": {
            "type": "boolean",
            "title": "Dryrunforcedbyenv"
          },
          "dryRunEnvVars": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Dryrunenvvars"
          },
          "providers": {
            "items": {
              "$ref": "#/components/schemas/ProviderOut"
            },
            "type": "array",
            "title": "Providers"
          },
          "resolvedDefault": {
            "$ref": "#/components/schemas/ResolvedDefaultOut"
          },
          "inheritedRoutes": {
            "type": "integer",
            "title": "Inheritedroutes"
          },
          "workspaceWritable": {
            "type": "boolean",
            "title": "Workspacewritable"
          }
        },
        "type": "object",
        "required": [
          "configPath",
          "configError",
          "model",
          "dryRun",
          "dryRunForcedByEnv",
          "dryRunEnvVars",
          "providers",
          "resolvedDefault",
          "inheritedRoutes",
          "workspaceWritable"
        ],
        "title": "SettingsResponse",
        "description": "The full ``GET /api/settings`` body."
      },
      "SettingsUpdateRequest": {
        "properties": {
          "model": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ModelInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "dryRun": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dryrun"
          },
          "clearModel": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clearmodel"
          }
        },
        "type": "object",
        "title": "SettingsUpdateRequest",
        "description": "Body of ``PUT /api/settings``.\n\n``None`` means \"leave this alone\", which is what lets the dry-run switch be\ntoggled without resubmitting the model form (and vice versa). Removing the\nin-app model entirely is the explicit ``clear_model`` flag rather than an\nambiguous ``model: null``."
      },
...
      "TestConnectionRequest": {
        "properties": {
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "baseUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseurl"
          },
          "apiKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apikey"
          }
        },
        "type": "object",
        "title": "TestConnectionRequest",
        "description": "Body of ``POST /api/settings/test`` -- all fields optional.\n\nAnything omitted falls back to the saved configuration, so the common\ncase (``{}``) tests exactly what a compile would use. Supplying fields is\nhow the screen tests an unsaved form."
      },
      "TestConnectionResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "latencyMs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latencyms"
          },
          "model": {
            "type": "string",
            "title": "Model"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "detail",
          "latencyMs",
          "model"
        ],
        "title": "TestConnectionResponse",
        "description": "The result of one test call.\n\nA provider refusing the key, timing out, or being unreachable are all\n*results*, not HTTP errors: the screen shows the detail inline next to the\nform the user needs to fix."
      },
...
    }
  }
}